+++ src/main/java/Model/ChanceCards/CardFactory.java
@@ -1,4 +1,4 @@
-package Model.Kort;
+package Model.ChanceCards;
 
 import java.awt.*;
