+++ src/main/java/Model/ChanceCards/FreePropertyCard.java
@@ -1,4 +1,4 @@
-package Model.Kort;
+package Model.ChanceCards;
 
 import Model.Player;
