+++ src/main/java/Model/Fields/JailField.java
@@ -1,15 +1,14 @@
-package Model.Felter;
+package Model.Fields;
 
-import Model.Field;
 import Model.Player;
 import gui_fields.GUI_Field;
-import gui_fields.GUI_Refuge;
+import gui_fields.GUI_Jail;
 
 import java.awt.*;
 
-public class FreeParkingField extends Field {
+public class JailField extends Field {
 
-    FreeParkingField(String name, String subText, String description) {
+    JailField(String name, String subText, String description) {
         super(name, subText, description);
     }
 
@@ -31,11 +30,11 @@ public class FreeParkingField extends Field {
 
     @Override
     public GUI_Field makeGUIFields() {
-        return new GUI_Refuge("default", this.getName(), this.getSubText(), this.getDescription(), Color.white, Color.black);
+        return new GUI_Jail("default", this.getName(), this.getSubText(),
+                this.getDescription(), Color.white, Color.BLACK);
     }
 
     @Override
     public void fieldAction(Player player) {
-        player.setLastAction(" - Har fået gratis parkering. Intet sker.");
     }
 }
